Added more thumbnail workers + queue per worker

+++ src/frontend/workers/thumbnailGenerator.worker.ts
@@ -2,6 +2,7 @@ import fse from 'fs-extra';
 
 import { getThumbnailPath, needsThumbnail } from '../utils';
 import { thumbnailType, thumbnailMaxSize } from 'src/config';
+import { IThumbnailMessage } from '../ThumbnailGeneration';
 
 const generateThumbnailData = async (filePath: string): Promise<ArrayBuffer | null> => {
   const inputBuffer = await fse.readFile(filePath);
@@ -45,12 +46,17 @@ const generateThumbnailData = async (filePath: string): Promise<ArrayBuffer | nu
 };
 
 const generateAndStoreThumbnail = async (filePath: string, thumbnailDirectory: string) => {
+  const thumbnailFilePath = getThumbnailPath(filePath, thumbnailDirectory);
+
+  // Could already exist: maybe generated in another worker, when use scrolls up/down repeatedly
+  // but this doesn't help if we want to deliberately overwrite the thumbnail, but we don't have that currently
+  if (await fse.pathExists(thumbnailFilePath)) {
+    return thumbnailFilePath;
+  }
+
   const thumbnailData = await generateThumbnailData(filePath);
   if (thumbnailData) {
-    const thumbnailFilePath = getThumbnailPath(filePath, thumbnailDirectory);
-
     await fse.outputFile(thumbnailFilePath, Buffer.from(thumbnailData));
-
     return thumbnailFilePath;
   }
   return '';
@@ -59,15 +65,39 @@ const generateAndStoreThumbnail = async (filePath: string, thumbnailDirectory: s
 // The worker context
 const ctx: Worker = self as any;
 
-// Respond to message from parent thread
-ctx.addEventListener('message', async (event) => {
-  const { filePath, thumbnailDirectory, fileId } = event.data;
+// Set up a queue of thumbnails that need to be processed
+// Without a queue, I've had to restart my computer since everything froze
+// (not 100% sure whether that was the cause)
+// TODO: Max queue length, so that when the user scrolls a lot, the most recent images will show up earlier?
+// (-> discard old requests)
+const queue: IThumbnailMessage[] = [];
+const MAX_PARALLEL_THUMBNAILS = 4; // Related to amount of workers. Currently 4 workers with 4 thumbs in parallel = 16 thumbs parallel total
+let curParallelThumbnails = 0;
+
+async function processMessage(data: IThumbnailMessage) {
+  const { filePath, thumbnailDirectory, fileId } = data;
   try {
-    const thumbnailPath = await generateAndStoreThumbnail(filePath, thumbnailDirectory);
-    ctx.postMessage({ fileId, thumbnailPath: thumbnailPath || filePath });
+    // console.log('Processing thumbnail message', { data, curParallelThumbnails, queue });
+    if (curParallelThumbnails < MAX_PARALLEL_THUMBNAILS) {
+      curParallelThumbnails++;
+      const thumbnailPath = await generateAndStoreThumbnail(filePath, thumbnailDirectory);
+      ctx.postMessage({ fileId, thumbnailPath: thumbnailPath || filePath });
+      curParallelThumbnails--;
+    } else {
+      queue.push(data);
+    }
   } catch (err) {
+    curParallelThumbnails--;
     throw { fileId, error: err };
   }
+  if (curParallelThumbnails < MAX_PARALLEL_THUMBNAILS && queue.length > 0) {
+    processMessage(queue.shift()!); // "pop" from the queue. First elements are at the start, so shift em
+  }
+}
+
+// Respond to message from parent thread
+ctx.addEventListener('message', async (event) => {
+  await processMessage(event.data);
 });
 
 // Make the file importable
